+++ bfabric/src/bfabric/rest/token_data.py
@@ -4,10 +4,18 @@ import asyncio
 import contextlib
 import urllib.parse
 from datetime import datetime
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Annotated, Any
 
 import httpx
-from pydantic import BaseModel, Field, SecretStr, ConfigDict
+from pydantic import (
+    BaseModel,
+    ConfigDict,
+    Field,
+    SecretStr,
+    ValidationInfo,
+    ValidatorFunctionWrapHandler,
+    WrapValidator,
+)
 
 from bfabric.entities.core.import_entity import import_entity
 
@@ -16,22 +24,40 @@ if TYPE_CHECKING:
     from bfabric.entities import Dataset, Instrument, Order, Plate, Project, Resource, Run, Sample, Workunit
 
 
+def parse_boolean_string(v: str, handler: ValidatorFunctionWrapHandler, info: ValidationInfo) -> bool:
+    """Parses a boolean string "true" or "false" to a boolean value."""
+    return {"true": True, "false": False}[v]
+
+
+BooleanString = Annotated[bool, WrapValidator(parse_boolean_string)]
+
+
 class TokenData(BaseModel):
     """Parsed token data from the B-Fabric token validation endpoint."""
 
     model_config = ConfigDict(populate_by_name=True, str_strip_whitespace=True)
 
     job_id: int = Field(alias="jobId")
+    """ID of the job associated with the token."""
     application_id: int = Field(alias="applicationId")
+    """ID of the B-Fabric application which created the token."""
 
     entity_class: str = Field(alias="entityClassName")
+    """Target entity class name"""
     entity_id: int = Field(alias="entityId")
+    """Target entity ID"""
 
     user: str = Field(alias="user")
+    """User/login"""
     user_ws_password: SecretStr = Field(alias="userWsPassword")
+    """Webservice password of the user."""
 
     token_expires: datetime = Field(alias="expiryDateTime")
+    """Expiration datetime of the token."""
+    web_service_user: BooleanString = Field(alias="webServiceUser")
+    """Indicates whether the user has permission to use webservices API"""
     caller: str
+    """The B-Fabric instance where the token originates from."""
     environment: str
 
     # Define a custom serializer method for model_dump