+++ bfabric/docs/changelog.md
@@ -9,6 +9,10 @@ Minor breaking changes are still possible in `1.X.Y` but we try to announce them
 
 ## \[Unreleased\]
 
+### Added
+
+- `TokenData.web_service_user` which indicates whether the user has permission to use webservices API.
+
 ## \[1.14.1\] - 2025-12-02
 
 ### Added